Increment plugin version to 1.6
Need to ensure this version is in use prior to putting out any newer nzbget "testing" builds as later builds (eg. r705) contain bugfixes to existing scripts in ppscripts directory and without the change to plugin code in 1.6, users upgrading to later testing versions will not get the newer changes to the scripts in ppscripts.